leftover item from previous PR: fixed unit tests in helm and fluxv2 plugins that had to do with GetInstalledPackageResourceRefs()
