+++ cmd/kubeapps-apis/plugins/helm/packages/v1alpha1/getresourcerefs_test.go
@@ -21,230 +21,67 @@ import (
 	"github.com/google/go-cmp/cmp/cmpopts"
 	"github.com/kubeapps/kubeapps/cmd/apprepository-controller/pkg/apis/apprepository/v1alpha1"
 	corev1 "github.com/kubeapps/kubeapps/cmd/kubeapps-apis/gen/core/packages/v1alpha1"
-	"github.com/kubeapps/kubeapps/cmd/kubeapps-apis/plugins/pkg/resourcerefs"
+	"github.com/kubeapps/kubeapps/cmd/kubeapps-apis/plugins/pkg/resourcerefs/resourcerefstest"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 type TestCase struct {
-	name               string
-	existingReleases   []resourcerefs.TestReleaseStub
+	baseTestCase       resourcerefstest.TestCase
 	request            *corev1.GetInstalledPackageResourceRefsRequest
 	expectedResponse   *corev1.GetInstalledPackageResourceRefsResponse
 	expectedStatusCode codes.Code
 }
 
+func newTestCase(tc int, identifier string, response bool, code codes.Code) TestCase {
+	newCase := TestCase{
+		baseTestCase: resourcerefstest.TestCases1[tc],
+		request: &corev1.GetInstalledPackageResourceRefsRequest{
+			InstalledPackageRef: &corev1.InstalledPackageReference{
+				Context: &corev1.Context{
+					Cluster:   "default",
+					Namespace: "default",
+				},
+				Identifier: identifier,
+			},
+		},
+	}
+	if response {
+		newCase.expectedResponse = &corev1.GetInstalledPackageResourceRefsResponse{
+			Context: &corev1.Context{
+				Cluster:   "default",
+				Namespace: "default",
+			},
+			ResourceRefs: resourcerefstest.TestCases1[tc].ExpectedResourceRefs,
+		}
+	}
+	newCase.expectedStatusCode = code
+	return newCase
+}
+
 func TestGetInstalledPackageResourceRefs(t *testing.T) {
 
-	// TODO (gfichtenholt) what's missing here is a call to
-	// resourcerefs_test.go init() fuction. I spent quite some time but have not yet
-	// figured out how to make that happen. So I am commenting this test out until I do
-
-	if len(resourcerefs.TestCases1) == 0 {
-		t.Logf("Expected non-empty array [resourcerefs.TestCases1]")
+	// sanity check
+	if len(resourcerefstest.TestCases1) < 11 {
+		t.Fatalf("Expected array [resourcerefstest.TestCases1] size of at least 11")
 	}
 
 	testCases := []TestCase{
-		/*
-				{
-					name:             resourcerefs.TestCases1[0].Name,
-					existingReleases: resourcerefs.TestCases1[0].ExistingReleases,
-					request: &corev1.GetInstalledPackageResourceRefsRequest{
-						InstalledPackageRef: &corev1.InstalledPackageReference{
-							Context: &corev1.Context{
-								Cluster:   "default",
-								Namespace: "default",
-							},
-							Identifier: "my-apache",
-						},
-					},
-					expectedResponse: &corev1.GetInstalledPackageResourceRefsResponse{
-						Context: &corev1.Context{
-							Cluster:   "default",
-							Namespace: "default",
-						},
-						ResourceRefs: resourcerefs.TestCases1[0].ExpectedResourceRefs,
-					},
-				},
-			}
-					, TestCase{
-						name:             tc.Name,
-						existingReleases: tc.ExistingReleases,
-						request: &corev1.GetInstalledPackageResourceRefsRequest{
-							InstalledPackageRef: &corev1.InstalledPackageReference{
-								Context: &corev1.Context{
-									Cluster:   "default",
-									Namespace: "default",
-								},
-								Identifier: "my-apache",
-							},
-						},
-						expectedResponse: &corev1.GetInstalledPackageResourceRefsResponse{
-							Context: &corev1.Context{
-								Cluster:   "default",
-								Namespace: "default",
-							},
-							ResourceRefs: tc.ExpectedResourceRefs,
-						},
-					}, TestCase{
-						name:             tc.Name,
-						existingReleases: tc.ExistingReleases,
-						request: &corev1.GetInstalledPackageResourceRefsRequest{
-							InstalledPackageRef: &corev1.InstalledPackageReference{
-								Context: &corev1.Context{
-									Cluster:   "default",
-									Namespace: "default",
-								},
-								Identifier: "my-apache",
-							},
-						},
-					}, TestCase{
-						name:             tc.Name,
-						existingReleases: tc.ExistingReleases,
-						request: &corev1.GetInstalledPackageResourceRefsRequest{
-							InstalledPackageRef: &corev1.InstalledPackageReference{
-								Context: &corev1.Context{
-									Cluster:   "default",
-									Namespace: "default",
-								},
-								Identifier: "my-apache",
-							},
-						},
-						expectedResponse: &corev1.GetInstalledPackageResourceRefsResponse{
-							Context: &corev1.Context{
-								Cluster:   "default",
-								Namespace: "default",
-							},
-							ResourceRefs: tc.ExpectedResourceRefs,
-						},
-					}, TestCase{
-						name:             tc.Name,
-						existingReleases: tc.ExistingReleases,
-						request: &corev1.GetInstalledPackageResourceRefsRequest{
-							InstalledPackageRef: &corev1.InstalledPackageReference{
-								Context: &corev1.Context{
-									Cluster:   "default",
-									Namespace: "default",
-								},
-								Identifier: "my-iis",
-							},
-						},
-						expectedStatusCode: codes.NotFound,
-					}, TestCase{
-						name:             tc.Name,
-						existingReleases: tc.ExistingReleases,
-						request: &corev1.GetInstalledPackageResourceRefsRequest{
-							InstalledPackageRef: &corev1.InstalledPackageReference{
-								Context: &corev1.Context{
-									Cluster:   "default",
-									Namespace: "default",
-								},
-								Identifier: "my-apache",
-							},
-						},
-						expectedStatusCode: codes.Internal,
-					}, TestCase{
-						name: tc.Name,
-						// See https://github.com/kubeapps/kubeapps/issues/632
-						existingReleases: tc.ExistingReleases,
-						request: &corev1.GetInstalledPackageResourceRefsRequest{
-							InstalledPackageRef: &corev1.InstalledPackageReference{
-								Context: &corev1.Context{
-									Cluster:   "default",
-									Namespace: "default",
-								},
-								Identifier: "my-apache",
-							},
-						},
-						expectedResponse: &corev1.GetInstalledPackageResourceRefsResponse{
-							Context: &corev1.Context{
-								Cluster:   "default",
-								Namespace: "default",
-							},
-							ResourceRefs: tc.ExpectedResourceRefs,
-						},
-					}, TestCase{
-						name:             tc.Name,
-						existingReleases: tc.ExistingReleases,
-						request: &corev1.GetInstalledPackageResourceRefsRequest{
-							InstalledPackageRef: &corev1.InstalledPackageReference{
-								Context: &corev1.Context{
-									Cluster:   "default",
-									Namespace: "default",
-								},
-								Identifier: "my-apache",
-							},
-						},
-						expectedResponse: &corev1.GetInstalledPackageResourceRefsResponse{
-							Context: &corev1.Context{
-								Cluster:   "default",
-								Namespace: "default",
-							},
-							ResourceRefs: tc.ExpectedResourceRefs,
-						},
-					}, TestCase{
-						name:             tc.Name,
-						existingReleases: tc.ExistingReleases,
-						request: &corev1.GetInstalledPackageResourceRefsRequest{
-							InstalledPackageRef: &corev1.InstalledPackageReference{
-								Context: &corev1.Context{
-									Cluster:   "default",
-									Namespace: "default",
-								},
-								Identifier: "my-apache",
-							},
-						},
-						expectedResponse: &corev1.GetInstalledPackageResourceRefsResponse{
-							Context: &corev1.Context{
-								Cluster:   "default",
-								Namespace: "default",
-							},
-							ResourceRefs: tc.ExpectedResourceRefs,
-						},
-					}, TestCase{
-						name: tc.Name,
-						// See https://kubernetes.io/docs/reference/kubernetes-api/authorization-resources/role-v1/#RoleList
-						existingReleases: tc.ExistingReleases,
-						request: &corev1.GetInstalledPackageResourceRefsRequest{
-							InstalledPackageRef: &corev1.InstalledPackageReference{
-								Context: &corev1.Context{
-									Cluster:   "default",
-									Namespace: "default",
-								},
-								Identifier: "my-apache",
-							},
-						},
-						expectedResponse: &corev1.GetInstalledPackageResourceRefsResponse{
-							Context: &corev1.Context{
-								Cluster:   "default",
-								Namespace: "default",
-							},
-							ResourceRefs: tc.ExpectedResourceRefs,
-						},
-					}, TestCase{
-						name: tc.Name,
-						// See https://kubernetes.io/docs/reference/kubernetes-api/authorization-resources/cluster-role-v1/#ClusterRoleList
-						existingReleases: tc.ExistingReleases,
-						request: &corev1.GetInstalledPackageResourceRefsRequest{
-							InstalledPackageRef: &corev1.InstalledPackageReference{
-								Context: &corev1.Context{
-									Cluster:   "default",
-									Namespace: "default",
-								},
-								Identifier: "my-apache",
-							},
-						},
-						expectedResponse: &corev1.GetInstalledPackageResourceRefsResponse{
-							Context: &corev1.Context{
-								Cluster:   "default",
-								Namespace: "default",
-							},
-							ResourceRefs: tc.ExpectedResourceRefs,
-						},
-					})
-				}
-		*/
+		newTestCase(0, "my-apache", true, codes.OK),
+		newTestCase(1, "my-apache", true, codes.OK),
+		newTestCase(2, "my-apache", true, codes.OK),
+		newTestCase(3, "my-apache", true, codes.OK),
+		newTestCase(4, "my-iis", false, codes.NotFound),
+		newTestCase(5, "my-apache", false, codes.Internal),
+		// See https://github.com/kubeapps/kubeapps/issues/632
+		newTestCase(6, "my-apache", true, codes.OK),
+		newTestCase(7, "my-apache", true, codes.OK),
+		newTestCase(8, "my-apache", true, codes.OK),
+		// See https://kubernetes.io/docs/reference/kubernetes-api/authorization-resources/role-v1/#RoleList
+		newTestCase(9, "my-apache", true, codes.OK),
+		// See https://kubernetes.io/docs/reference/kubernetes-api/authorization-resources/cluster-role-v1/#ClusterRoleList
+		newTestCase(10, "my-apache", true, codes.OK),
 	}
 
 	ignoredFields := cmpopts.IgnoreUnexported(
@@ -254,12 +91,12 @@ func TestGetInstalledPackageResourceRefs(t *testing.T) {
 	)
 
 	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
+		t.Run(tc.baseTestCase.Name, func(t *testing.T) {
 			authorized := true
 			actionConfig := newActionConfigFixture(
 				t,
 				tc.request.GetInstalledPackageRef().GetContext().GetNamespace(),
-				toHelmReleaseStubs(tc.existingReleases),
+				toHelmReleaseStubs(tc.baseTestCase.ExistingReleases),
 				nil)
 
 			server, _, cleanup := makeServer(t, authorized, actionConfig, &v1alpha1.AppRepository{
@@ -283,7 +120,7 @@ func TestGetInstalledPackageResourceRefs(t *testing.T) {
 	}
 }
 
-func toHelmReleaseStubs(in []resourcerefs.TestReleaseStub) []releaseStub {
+func toHelmReleaseStubs(in []resourcerefstest.TestReleaseStub) []releaseStub {
 	out := []releaseStub{}
 	for _, r := range in {
 		out = append(out, releaseStub{name: r.Name, namespace: r.Namespace, manifest: r.Manifest})
